rabbit_api/mq/requirements.txt
pika==1.3.1
python-dotenv==0.19.2